+++ grid_map_pcl/include/grid_map_pcl/grid_map_pcl.hpp
@@ -1,0 +1,12 @@
+/*
+ * grid_map_pcl.hpp
+ *
+ *  Created on: Oct 19, 2016
+ *      Author: Dominic Jud
+ *	 Institute: ETH Zurich, Robotic Systems Lab
+ */
+
+#pragma once
+
+#include <grid_map_core/grid_map_core.hpp>
+#include <grid_map_pcl/GridMapPclConverter.hpp>